refactor: group update tests
